Move checks for &Attribute-Name into radius_parse_attr()
and clean up the rest of the code in the server which checked
for '&'.

As a result, the rest of the code gets simpler.  We get better
error messages, and a few corner cases get fixed.  i.e. where
the '&' was NOT being handled, and radius_parse_attr() was being
called, with what SHOULD have been an attribute.  So if the poor
users followed the docs and did &Attribute-Name, they would get
a parse error, and their brains would explode.  We don't want
that...