Fix handling of proxying to virtual servers.  Fixes #596
The downside is that there isn't really any parent/child
relationship between the real request and the fake one.
So if the fake one blocks for 60s, the real one will be told
to stop, and the fake one with happily continue.

We'll need a separate commit to fix that issue.