Anything which eventually has the value it got interpreted by pairparsevalue, or doesn't pass the length of the string around must not pre-unescape the string
This broke pretty much all escape sequences in preprocess, sql, ldap and radclient. Arguably as conffile.c is the only place which needs this, it should probably be the thing doing the unescaping.

No, I don't buy that things should be prefixed with \\, that's shit. The server should be able to reingest what it spits out in detail files and there's no way that worked in the current server with octal escape sequences.