add a third query to start / stop
The queries have 'AND AcctStopTime IS NULL', which speeds up
the normal case substantially.  However, we can still have
the case where we get a delayed stop, or where the detail file
is being replayed multiple times.  We therefore want to be able
to insert the record where AcctStopTime is NOT NULL