3.0.x - Make EAP-SIM work again - proper encoding of EAP-SIM attributes within EAP-Message
This fix follows the issue I logged (on the mailing list, not on GitHub)
on June 11th.

As a reminder, the problem happened after a commit which (among other
things) modified the EAP-SIM attributes.
Since this commit, EAP-SIM authentication do not work because
EAP-Message is not properly encoded anymore by FreeRADIUS.

I believe the commit is the following:

https://github.com/FreeRADIUS/freeradius-server/commit/39df09e42d80a96363be0bddee2ff0ba97fdb035

So, here is a fix.

I also fixed the attributes issue in radeapclient, but at the moment the
binary is unusable: it crashes, and I don't have time to look into this.
(I tested the fix with another EAP client)