Move suppression of debugging messages to a better place.  Closes #772
vradlog() should always log.  It's static to log.c, so the *callers*
should take care to avoid calling vradlog().

The checks for debugging messages are pushed to radlog(), which
is the external API.

Added a static radlog_always(), which always calls vradlog().
It's just a wrapper to deal with the varargs stuff.

The vradlog_request() function now calls radlog_always(), as
vradlog_request() takes care of checking if debugging is enabled.