Change logging in radius_exec_program to use R* macros
RERROR should always be used where an error may indicate a failure of an external resource like a script, or database and we have the REQUEST available.

It outputs to the main server log like ERROR but also includes the request number, so multiple errors can be tied together when inspecting the log.