Fixed check in pairvalidate
The only commit in this series that actually changes some behaviour. The test was supposed to check if one of the values (match and check) was false, while the other was true. Because the while loop already validated that at least one of them is true, we can get the desired behaviour by just checking if one of them is false. The old behaviour only tested to see if match was false.

This code is not tested in the unit tests. It would be nice to have it in there, but I've got no idea how to write that.

This fixes the following message of cppcheck:

[src/lib/valuepair.c:547]: (style) Same expression on both sides of '||'.