Don't do \n --> CR mashing in xlat.c
As it turns out, it's almost always wrong.  The correct escaping
rules are that the conf file parser reads a string, and that string
is passed verbatim through all of the layers to the final consumer.

Only the final consumer does the mashing of \n --> CR.

This means that the tests are currently broken.

%{md5:...\n} expects to get a CR on input, and it now gets a "\n".
Fixing that is the next stage.