Expose core library versions and features in ${feature.*} and ${version.*}
This is mainly needed so that the regex flavour is available for shipped policies, but can also be used for the canonical store of module dependency version numbers.

That in itself isn't useful for the server config, as we'd need some sort of 'pass2' for the config sections, but it is useful for querying via radmin.