Address (some) potential TOCTOU attack vectors in rad_mkdir
We now leave ownership as the effective UID (which should be sufficiently restrictive) whilst creating the directory structure, and set initial permissions to 0700.

When setting final permissions, we first open the directory, check we still have write permissions by doing fchmod with our effective UID/GID (which should not be super user), then suid_up, and use fchown to set the owner.