Add channels to radmin.
there is now stdin, stdout, and stderr for radmin.
radmin writes to the server on stdin.
The server responds with data on stdout or stderr.
the server also responds with a status of the command

So that errors from the server go to the right place,
and radmin can exit with the status of the commands