Don't wait for a reply to DHCP release and decline
DHCP release and decline never get a reply, so don't wait for one.

Also I defined the requests types in a FR_NAME_NUMBER array (as is done
in radclient).
And I removed the "offer" command (because this is a response sent back
from the server, not something a client sends).