TCP socket timer limits mean they're marked frozen, not EOL
EOL is now only for errors.  a frozen socket can still have
requests using it.  For auth/acct, we can't close the socket
until all requests using it have finished.  We can't close
socket while the client is still sending packets.  That's rude.

For proxy sockets, most of the same applies.  The only difference
is that the frozen socket won't be used to proxy new requests.

But retransmits will still go out the proxy socket.  And replies
will still be received from it.