radeapclient - rate limit option
This is the first of a new set of commits to radeapclient.
If you're still ok with this, we can proceed the same way as for last
pull request:
I'll post small commits so you can see what's being changed; and at some
point we'll agree on a merge.

This commit adds option "-n" which allows to control the rate of
requests sent per second.