Fix for redundant-load-balance.  Closes #945
In normal operations, modcall_child / modcall_recurse processes
the current node, and all of its children.  For redundant-load-balance,
we want to loop BACK from the end of the list to the start, AND
stop when we reach the first one we found again.

This means we have to tell the functions "process ONE node only",
and do all "next" operations ourselves.