radeapclient - check response with filter
When receiving a response for which a filter is defined, use it (same as
in radclient).
The response packet code is always implicitly checked, even if no filter
is set (ditto).
Also added statistics "Failed filter", "Passed filter".

Note: the filter is only checked on the final response. (this means
that, for EAP transactions, there is no check on challenges).