Add "bootstrap" method, and clean up module API.
The bootstrap method is where attributes are created at run time.
This allows OTHER modules to use those attributes.

All modules in the "modules" section are now dlopen'd, their
instance data allocated, configuration is parsed, and the bootstrap
method is called.

Once all of the modules are bootstrapped, the module "instantiate"
method is called.

As part of this change, we simplify the API to:

	module_find = find a known module
	module_bootstrap = load and create a known module
	module_instantiate - call the modules instantiation routine