Simplify map_proc interface
In almost all cases instance data passed to escape functions is the module instance.
In all cases map_proc contexts should be hung off the module that registered them.
Simpler to use a similar instantiation system as we do for modules, where the memory is allocated by the server, then passed to the instantiate callback.