Move rlm_cache to using maps instead of VALUE_PAIRs
This greatly simplifies the cache code as it doesn't have to deal with multiple lists in the cache entry struct. It also exposes the full range of lists, and destinations e.g. outer. and coa.

Additionally, all operators that are allowed in update sections are now allowed in lists, including filter operators.