Globbing across directories is not supported.  Fixes #1149.
The rest of the "filename_work" code assumes that the detail.work
file is in the same directory as the files being globbed.  In order
to support globbing for directories, the code will have to be updated
to *not* open detail.work, but instead to do a *second* file
globbing() on .../*/detail.work, in order to get the real filename
of the detail.work file.

In the short term it's easier to just enforce the assumptions
in the rest of the code.