Start of "write normalized config"
Which requires WITH_CONF_WRITE to be defined manually for now.

The idea is for the server to be able to read a config file,
and then to write it out with comments, but normalized for tabs,
conditions, "update" sections, etc.

The normalized config can then be used as the default config,
which is better for "look and feel".  The normalized config
can also be used as the basis for automatically generated
documentation, too.