Add fr_pair_list_mcopy_by_num()
Which is like fr_pair_list_move(), but does copy / delete
instead of talloc_steal.

The problem is that talloc_steal() keeps the original parent
context around for the lifetime of the VP being stolen.  Which is
bad when the VP comes from a REQUEST, and is put into another
context, which lives for multiple seconds.