Move EAP to using only the new state API
Need to test proxying inner methods (and add tests) as this functionality may be broken.

We now use the state API to automagically persist and restore request_data_t that have been marked with the persist flag (this will work for anything, not just EAP).

For fake child requests we use the same session index, capturing it from the fake request after it returns, and setting it in the next fake request.  This avoids using the state tree for inner requests, and is fairly self contained, so isn't too hacky.

The first time we get a child eap_session, its lifetime is bound to its parent.  If the parent gets freed then the child will be also.  This probably won't happen ever, except possibly if the supplicant started sending malformed outer EAP packets after the tunnel had been setup.