Replaced array function parameters with pointers
This was already done implicitely, but writing it as an array suggests that it is passed by value instead, and that you can call things like `sizeof` on it. Some explanation and arguments can be found at https://lkml.org/lkml/2015/9/3/428 (although you have to read between the rants, as usual with texts of Linus).