Update policy.d/filter
Rename the filter_inner_identity implementation to filter_inner_outer_identity and...

1) Make the check for anonymous absolute and deterministic by looking for anonymous or anon, not just any stripped user-name that begins with anon. This is incorrect, poor behaviour as it would match any genuine identities that begin with anon.

This change should be acceptable and is sensible and reasonable, on balance, as the changes being made here are to the default configuration only, they do not apply to existing installations so there is no risk of breaking a current deployment. The behaviour could be modified as a new installation is configured if this ever became necessary.

Additionally, without doing this, we are not going to see feedback about what this would break.

A strict, absolute default is desirable as there is no ambiguity in what is considered an anonymized identity and what is not.

2) This strict definition then allows us to require the inner identity to not be anonymized, having defined that anonymous and anon are anonymized.

3) Change the logic to not allow subrealms in the inner identity where the outer identity is not anonymized: bob@secure.example.com (inner) and bob@example.com (outer) should fail. Presently, this passes.

4) Make the module failure messages more descriptive.