Hold on to your hats.  Proxies are now children.
This changes moves "request->proxy" from being a RADIUS_PACKET
to being a REQUEST.  This means one more set of dereferencing
to get request->proxy->packet, which is bad.  But it's 2016,
and CPUs are fast.

The benefit is that it's now possible to separate the state
machines for incoming requests, and outgoing requests. It also
means that we can now have different protocols for incoming
and outgoing packets.

This change will make new features much easier to implement.