Remove support for setting a default inner EAP method in EAP-TTLS
This seems vestigial and actively breaks things

We always require an inner-tunnel virtual server

The contents of the inner tunnel are essentially the body of a diameter packet.  If we do EAP, then the inner tunnel EAP module, processing the EAP-Message attributes should take care of negotiating the inner EAP type.