raddb: move driver-specific config out from mods-available
Symlinking them into mods-enabled, the location INCLUDEd in mods-available/sql,
breaks the server as it tries to load rlm_<driver>.so, which doesn't exist.
They're only part of a module config, so don't really fit that well in
mods-available.

Additionally, turn the $INCLUDE into an $-INCLUDE to pull the config in if it
exists, rather than having to be locally uncommented. This makes it easier to
configure sql for the first time (especially to get up and running with sqlite).