Use old PRF functions if SSL_export_keying_material() fails.
Because https://tools.ietf.org/html/rfc5705#section-6

says that certain labels are prohibited, and OpenSSL enforces
that prohibition.

We need the keys for EAP, so we have to *always* implement the
functions.  Because SSL_export_keying_material() might fail.

We'll need a later commit to implement the TLS 1.2 PRF.