Typo fix: Lenth => Length
This occurence was missed in commit a25ebeb3d7e414c51013fdee476ba4886399f9a3:
That commit was backported from v3.0.x and rlm_json is new in v3.1.x.

We will make a few changes to this method and copy-paste the doxygen
comments, so clean them up as soon as possible.