Update memory handling in rlm_json
JSON-C includes its own memory management system, which uses a
refcounting system. In this case, the difference between `free` and
`json_object_put` were none, but it might serve as a bad reference to
include the call to `free` here.

Another change: the successful code path completely forgot to free the
temporary object. This asked for a little bit of restructuring, which
removed the goto lines.