Add support for systemd's READY notification.
When starting a unit systemd is able to wait for a notification from the
process indicating a successful start, rather than assuming a successful start
as soon as it has forked.

Upon daemonizing the exiting parent will emit a READY=1 notification once the
child has indicated successful initialisation.

Use the following in the unit file to rely on the notificaton.

[Service]
Type=notify

Note: systemd support is compiled out by default, enabled with --with-systemd.
May want to change this in future.