Add support for systemd watchdog notifications.
systemd is able to monitor a process to ensure that it periodically sends
watchdog notifications.

We install a watchdog event into the event loop that emits the notification and
reschedules the next event in half the watchdog interval provided by systemd.

To enable watchdogs add something like this to the unit file:

[Service]
WatchdogSec=10
NotifyAccess=all

Reliably detecting that systemd requires watchdog notifications at a certain
interval must be performed in the initial process, however the subsequent
notification can be emitted by any child.

At present this simply ensures that the event loop is running.