Catch broken configurations.
Version 4 is composed of many new subsystems.  Some like async
unlang interpreter are pretty much done.  Others, like the new
virtual server sections, are still in progress.

If a new subsystem is UNDOCUMENTED, and has NO EXAMPLE CONFIGURATION,
you *SHOULD NOT BE USING IT*.

Bug reports will be closed as "not relevant".  The subsystem is
partially done, and may not even work from day to day.  Posting
bug reports for code that is being worked on is almost always
unproductive.