Fixed compilation problems with rlm_lua
When building the debian package, no configuration is tried for rlm_lua,
because it isn't listed in the stable modules. But because the folder
contains a file called all.mk, the build process still tries to build it
and fails. Quick and dirty solution: move all.mk to all.mk.in.