update documentation to match code behavior.
v4 now only has *filtering*, not *enforcement* operators.  That is,
if "Session-Timeout = 0", doing:

	update request {
		Session-Timeout > 86400
	}

Will *delete* the non-matching Session-Timeout.  It will not *add*
or *update* the Session-Timeout to be 86400.

Which means that in v4, we should probably have a "filter" section,
instead of an "update" section, to make this behavior clearer.