Add fr_dict_attr_add_by_name().  Addresses #1883
This function adds an attribute ONLY by name, and not by number.
It can be used to add Attr-26, without polluting Vendor-Specific
for everyone else.

The dictionary code assumes that when two attributes are defined
for the same number, that the second one takes precedence over
the first one.  See "dictionary.compat" for examples