Simplify unknown attribute OID parsing/printing
As tmpls are no longer stack allocated, we can hang additional talloced memory off the structure.

This means unknown attributes no longer need to be placed in a buffer which is part of the tmpl, but can be allocated on the heap.

It also means that where an unknown attribute hierarchy contains an unknown vendor, the unknown vendor can form part of the hierarchy too, without needing a special field.

References #1883