Fixed parameters order in fr_pair_find_by_num()
This commit fixes the parameters passing order in function
fr_pair_find_by_num().

In the original version, the variable hwvp was always NULL due to a
failure of function fr_pair_find_by_num() caused by the bad order of
parameters. This bug avoids the update of the ARP table when DHCP offers
are sent unicast.