Add --disable-openssl-compat-check configure option
Add support for --disable-openssl-compat-check configure option, which
disables checking whether the linked OpenSSL (or related) libraries are
compatible with the ones FreeRADIUS was built with.

This is needed for distribution, which are tracking OpenSSL ABI changes
themselves and make sure shared library major and minor numbers change
if there is any breakage, thus ensuring incompatible applications and
libraries are not linked. That is also supported by the new OpenSSL
upstream policy to keep ABI compatible between "fix" and "patch"
version numbers.