Start the tedious work of splitting evaluation of config structures from parsing them
This should allow multiple front ends for getting config info.  There could be one which converts LDAP objects to CONF_SECTIONs and CONF_PAIRs for example.

The real reason for this however is to start teasing out evaluation of conditions and maps from parsing of the config files.  It needs to happen separately in a different phase to parsing, once we've discovered what protocol the virtual server is using.

Before the evaluation stage the server only needs to understand the basic grammar of sections (with and without arguments), and conf pairs.  Even a conditional statement is just a section with two arguments (one of which happens to be quoted with parenthesies).