updated ldap default config
net_timeout has a default setting of 10 in the code but the enabled value in the default config is 1 - therefore
the 'default' would be , for users, 1 - so commented out. users can enable it to change.
likewise, same for ldap_debug - 'default' is 0x0000 but it had a preconfigured value of 0x0028 - so commented
out so default is in play and admins can choose to change it.