Define home server expiration field and configuration option.
TLS keys associated to dynamic home servers and realms expire.
However, FR does not know anything about this expiration time.
Instead, FR tries to use a home server until it fail to establish
the connection. Then a new REALM is negotiated and used until
the new keys expire.

This behaviour has the problem that it requires a failed connection
(and hence authentication) to realize REALM's keys are no longer
usable.

This commit introduces a new filed into the home_server_t struct,
called "expiration" that marks the point in time after the REALM
should not be used.

Ideally this field is established based on the received information
from libtr. However, an accessor for this information is missing.
While this accessor is implemented, a new configuration option
"realm_lifetime" is provided to manually indicate the lifetime of
a realm.

NOTE: Whereas "home_server_t.expiration" is an absolute time_t value
indicating a timestamp, "realm_lifetime" is a relative amount
of time to be added to time() at a REALM creation.