Implemented rekey functionality for the trust router client.
* Added rekeyer thread running a specific event list.
* Added option to enable/disable rekey functionality.

Each time a REALM is successfully queried from the trust router server, a
rekey event is scheduled to be executed 120s before the TLS keys associated
to that REALM expire (based on the home_server_t.expiration filed).

A rekey event consists on executing a new TIDC query using the same parameters
as for the original creation of the dynamic REALM, obtaining a fresh set of home
servers with new TLS keys.

If the rekey process succeeds, a new rekey event is scheduled as described
above. If a rekey event fails, it tries again after 10 seconds.
If it fails 5 times in a row, it stops trying to rekey the realm.