use RLM_MODULE_UNKNOWN for unset codes
this separates the default return code from the one currently
being processed.

i.e. we push the default code onto the stack, with an empty
instruction and the top_frame flag set.  We then push the real
instruction, along with UNKNOWN rcode.  That way, if the unlang
is all if / elsif (which doesn't change priorities), we can
keep going, and not stop on the default return code.

Once the section has been processed, it will merge in the
top return code.