remove exfile_unlock()
It was only used in the detail file, when the detail file
called fdopen().  The exfile API called dup() just so that the
detail module could call fclose() on the fdopen'd FILE pointer,
and not have the underlying FD close.

It's better to just have the detail module dup() the FD itself,
and leave the main FD unchanged