suppress multiple sleeping signals.
If we're running single-threaded. then the network && worker
are in the same event list.  So the worker sending a signal
saying "I'm sleeping" will cause the event loop to get run
again, including the worker pre-handler, which signals the
event loop that it's sleeping.

It's best to remove infinite loops.