create and use Session-State-User-Name
so that we can cache the User-Name in Session-State without
affecting the User-Name in the reply.

The default "send Access-Accept" policy adds all of the session-state
attributes to the reply.  Which in this case would mean User-Name...
but we don't want that.  i.e. we don't want our internal policies
to pollute whatever the user wants to send back