different way of hacking const issues
in non-debug bilds, talloc_get_type_abort() is just a cast.
but if the variable is 'const', it's talloc_type is the *non-const*
version.  So we can't just add 'const' to the arguments to
talloc_get_type_abort().  That way the debug builds would fail...

the solution seems to be jist removing the calls entirely in
non-debug builds.