allow adding "watch directory" to the network side
the function called from the event loop is empty and does
nothing for now.  Later, it has to call a new function in
the app_io, and tell it "something changed in the directory".

That function then needs to open the new detail file, and
(somehow?) get those packets back to the fr_network_t, for
forwarding to a worker.