EAP-FAST: Fix EAP-FAST-MSCHAPv2 in anonymous tunnel (v3.0.x)
According to RFC 5422, in anonymous FAST tunnel the mschapv2
challenge is generated differently, from the phase1 exchange.
While the eap-fast module takes care to save internally the
peer-challenge to use, no one cares to read it before doing
the mschap check, causing it to fail.
Fix this by checking for PW_MS_CHAP_PEER_CHALLENGE and using it.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>