Ensure request is always available at FR_TLS_EX_INDEX_REQUEST
One of the reasons why this code was rewritten for v3.1.x/v4.0.x we because it didn't seem like request was always available in the correct places. This at least ensures that request is available in any TLS callbacks triggered by the EAP methods even if it isn't very pretty.