keep timers around even when removing packets from a connection
so that the timers keep running, and the packets are retransmitted
when the timers hit.

This shouldn't be an issue normally. If a new connection opens,
any pending t->queued are drained, and sent immediately. BUT
there may be some pending packets in the queue which didn't make
it to the connection.  So we have to handle that case, too.

If there is no connection, any timed-out packets are sent only when
a connection is available.  This means that the timers aren't
perfect in that case, but there's no perfect situation here.