encode temporary packet, not request->packet.  Address #2092
We can't call fr_dchp_encode(request->packet) because
request->packet->data != NULL.  It's also bad form to muck with
the incoming packet.

The solution is to create a temporary packet, and then encode
and send that one.

Also, dhcprelay_process_server_reply() poked request->packet
in some places, and request->reply in others.  So it was impossible
for it to work.