clean up schedule exit code
because of back & forth signaling issues, we have to be careful
how we clean up memory.

First, the network is signaled to exit.  When it does, the
scheduler thread sends a semaphore saying so.  That signals the
master thread to call fr_network_destroy(), which in turn signals
the worker threads that it's exiting.

BUT fr_network_destory() doesn't free underlying memory.  Instead,
we have to wait for all of the workers to finish.

The workers are signaled to finish, and again don't call
fr_worker_destroy().  Instead, they don't ACK the channel close,
but do signal the semaphore again.

When all of the workers are done, it's safe to free the network
memory.  And then it's also safe to free the worker memory.

... and then disable most of it, because apparently even
signaling the workers to exit causes problems.