Ensure that libkqueue / Linux doesn't bork us.
The kqueue API takes an FD, but inotify requires a filename.
libkqueue uses /proc/PID/fd/# to look up the FD -> filename mapping.

However, if you start the server as "root", and then swap to "radiusd",
/proc/PID will be owned by "root" for security reasons.  The only way
to make /proc/PID owned by "radiusd" is to set the DUMPABLE flag.

Instead of making the poor sysadmin figure this out,
we check for this situation, and give them a
descriptive message telling them what to do.