Use EXIT_FAILURE for portability
Apparently not all systems use 1 to indicate failure.  C99 says "Finally, control is returned to the host environment. If the value of status is zero or EXIT_SUCCESS, an implementation-deﬁned form of the status successful termination is returned. If the value of status is EXIT_FAILURE, an implementation-deﬁned form of the status unsuccessful termination is returned. Otherwise the status returned is implementation-deﬁned."