Reorganise attribute adding/validation functions and add support for internal dictionary references.
References allow one attribute to internally reference another TLV or a dictionary root.

This allows us to create "group" attributes which allow any attribute to be encoded within the group and foreign protocol attributes which allow RADIUS to be encoded within DHCP (for example).

Attribute references are also needed to mix protocols in the same virtual server, as the template validation functions can retrieve the foreign dictionary context from the talloc hierarchy.