Do not encode htype and hlen by default for DHCPv4
DHCP fields "htype" and "hlen" should not be encoded to 1 and 6 (respectively) by default.
They don't need to, because they are automatically set if there is an Ethernet address in field "chaddr".

For a Lease Query request, as described in RFC 4388, it is possible to query: 1) by IP address, 2) by MAC address, 3) by Client-identifier option.
In cases 1) and 3) "The values of htype, hlen, and chaddr MUST be set to zero.".

So I think that if no "chaddr" is provided, it's better to leave "htype" and "hlen" to 0.

This proposed pull request implements this change.