If values need to be zero padded they should be fed into the padding expansions
It's not ok to mess with the underlying data type just to get a particular presentation format.

The whole point of reworking the xlat code to take discreet boxed values is that upstream consumers of the output can use real types to make proper casting and comparison decisions.

By forcing these values to strings there's a high chance that they'll be fed into a lexicographic comparison function instead of the server just using integer comparisons (they would be now anyway, but we're moving towards something better).