iline number is not order.
The "users" file may have multiple $INCLUDES.
We DON'T want a "user" entry at line 266 of "$INCLUDE 1" to conflict
with a "DEFAULT" entry at line 266 of "$INCLUDE 2".

So we add an "order", which tracks the entry number.  The "lineno"
is no longer used for ordering comparisons.  Instead, it's only
used for debugging.

Manual port of commit bee85afcc61