fix wildcard includes
so that you can now do:

    $INCLUDE mods-enabled/sql
    $INCLUDE mods-enabled/

This will cause the "sql" module to be read first.  Then, when the
directory is read, the "sql" module will be ignored.

This solves ordering issues where "sqlippool" needs "sql", but
"sqlippool" is listed in the *directory* on disk after "sql".

Manual port of commit bb842ece0102