start of separation of config && thread data for network IO
we now have "app_io_instance" which is the parsed configuration
data.

And we also have "socket_instance", which is the data that changes
at run-time.  e.g. socket FD, list of dynamic clients, statistics,
etc.

This commit just adds the API capabilities. The next few commits
should start separating the responsibilities of the various pieces
of information.